Fix startup with theme from preferences.
- DynamicThemeViewModel handles preferences.
- Show progress indicator on initial load, show error on failure
- On detail screen, show selected source color and variant.

+++ app/src/main/kotlin/com/github/mheerwaarden/dynamictheme/ui/DynamicThemeViewModel.kt
@@ -15,46 +15,105 @@
  * limitations under the License.
  */
 
-package com.github.mheerwaarden.dynamictheme.ui.screen
+package com.github.mheerwaarden.dynamictheme.ui
 
 import android.util.Log
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.mheerwaarden.dynamictheme.APP_TAG
 import com.github.mheerwaarden.dynamictheme.data.database.DynamicTheme
 import com.github.mheerwaarden.dynamictheme.data.database.DynamicThemeRepository
 import com.github.mheerwaarden.dynamictheme.data.database.Id
-import com.github.mheerwaarden.dynamictheme.data.preferences.UserPreferences
+import com.github.mheerwaarden.dynamictheme.data.preferences.UserPreferencesRepository
 import com.github.mheerwaarden.dynamictheme.material.color.utils.ColorExtractor
-import com.github.mheerwaarden.dynamictheme.ui.ColorSchemeState
-import com.github.mheerwaarden.dynamictheme.ui.screen.DynamicThemeUiState.Companion.createDynamicThemeUiState
-import com.github.mheerwaarden.dynamictheme.ui.screen.DynamicThemeUiState.Companion.fromDynamicTheme
+import com.github.mheerwaarden.dynamictheme.ui.DynamicThemeUiState.Companion.createDynamicThemeUiState
+import com.github.mheerwaarden.dynamictheme.ui.DynamicThemeUiState.Companion.fromDynamicTheme
+import com.github.mheerwaarden.dynamictheme.ui.screen.UiColorSchemeVariant
 import com.github.mheerwaarden.dynamictheme.ui.theme.DarkColorScheme
 import com.github.mheerwaarden.dynamictheme.ui.theme.LightColorScheme
-import com.github.mheerwaarden.dynamictheme.ui.toColorSchemeState
 import dynamiccolor.Variant
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.time.LocalDateTime
 
 private const val TAG = APP_TAG + "_DynamicThemeViewModel"
 
+sealed interface LoadingState {
+    data object Loading : LoadingState
+    data object Success : LoadingState
+    data class Failure(val error: Throwable) : LoadingState
+}
+
 /**
  * A view model that is used in all the steps to select a color and a scheme variant
  */
 open class DynamicThemeViewModel(
     private val dynamicThemeRepository: DynamicThemeRepository,
+    private val userPreferencesRepository: UserPreferencesRepository,
+    private val isPreferenceState: Boolean = true,
     var onException: (String) -> Unit = { _ -> },
 ) : ViewModel() {
 
     protected val _uiState = MutableStateFlow(DynamicThemeUiState())
     val uiState: StateFlow<DynamicThemeUiState> = _uiState.asStateFlow()
+
+    var loadingState: LoadingState by mutableStateOf(LoadingState.Loading)
+
+    init {
+        updateFromPreferences()
+    }
+
+    fun updateFromPreferences() {
+        loadingState = LoadingState.Loading
+        viewModelScope.launch {
+            try {
+                withContext(Dispatchers.IO) {
+                    val preferences = userPreferencesRepository.preferences.first()
+                    _uiState.value = if (preferences.id < 0) {
+                        // Initialize from preference values
+                        createDynamicThemeUiState(
+                            id = preferences.id,
+                            name = preferences.name,
+                            sourceColorArgb = preferences.sourceColor,
+                            uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(preferences.dynamicSchemeVariant),
+                            windowWidthSizeClass = _uiState.value.windowWidthSizeClass
+                        )
+                    } else {
+                        // Initialize from database values
+                        val theme = dynamicThemeRepository.getDynamicTheme(preferences.id)
+                        if (theme == null) {
+                            createDynamicThemeUiState(
+                                id = -1,
+                                name = preferences.name,
+                                sourceColorArgb = preferences.sourceColor,
+                                uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(preferences.dynamicSchemeVariant),
+                                windowWidthSizeClass = _uiState.value.windowWidthSizeClass
+                            )
+                        } else {
+                            fromDynamicTheme(theme)
+                        }
+                    }
+                }
+                loadingState = LoadingState.Success
+            } catch (e: Exception) {
+                Log.e(TAG, "updateFromPreferences: Exception during update: ${e.message}")
+                loadingState = LoadingState.Failure(e)
+            }
+        }
+    }
 
     fun updateWindowSizeClass(newWindowSizeClass: WindowSizeClass) {
         // Update only after reconfiguration
@@ -66,6 +125,9 @@ open class DynamicThemeViewModel(
 
     fun updateName(newName: String) {
         _uiState.value = _uiState.value.copy(name = newName)
+        if (isPreferenceState) {
+            setNamePreference(newName)
+        }
     }
 
     fun updateColorScheme(
@@ -76,38 +138,13 @@ open class DynamicThemeViewModel(
             id = _uiState.value.id,
             name = _uiState.value.name,
             sourceColorArgb = sourceColorArgb,
-            uiColorSchemeVariant = uiColorSchemeVariant
+            uiColorSchemeVariant = uiColorSchemeVariant,
+            windowWidthSizeClass = _uiState.value.windowWidthSizeClass
         )
-    }
-
-    fun updateFromPreferences(preferences: UserPreferences) {
-        if (preferences.id < 0) {
-            // Initialize from preference values
-            _uiState.value = createDynamicThemeUiState(
-                id = preferences.id,
-                name = preferences.name,
-                sourceColorArgb = preferences.sourceColor,
-                uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(preferences.dynamicSchemeVariant)
+        if (isPreferenceState) {
+            setSourceColorPreference(
+                sourceColorArgb, uiColorSchemeVariant.toVariant()
             )
-        } else {
-            // Initialize from database values
-            viewModelScope.launch {
-                val dynamicThemeUiState = withContext(Dispatchers.IO) {
-                    val theme = dynamicThemeRepository.getDynamicTheme(preferences.id)
-                    if (theme == null) {
-                        createDynamicThemeUiState(
-                            id = -1,
-                            name = preferences.name,
-                            sourceColorArgb = preferences.sourceColor,
-                            uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(preferences.dynamicSchemeVariant)
-                        )
-                    } else {
-                        fromDynamicTheme(theme)
-                    }
-                }
-                _uiState.value = dynamicThemeUiState
-            }
-
         }
     }
 
@@ -127,6 +164,9 @@ open class DynamicThemeViewModel(
                 val newId =
                         dynamicThemeRepository.insertDynamicTheme(_uiState.value.toDynamicTheme())
                 _uiState.value = _uiState.value.copy(id = newId)
+                if (isPreferenceState) {
+                    setIdPreference(newId)
+                }
                 Log.d(TAG, "insertDynamicTheme: DynamicTheme added: $newId")
             } catch (e: Exception) {
                 val msg = "addDynamicTheme: Exception during insert: ${e.message}"
@@ -155,14 +195,41 @@ open class DynamicThemeViewModel(
 
     /** Delete the dynamicTheme from the database. */
     fun deleteDynamicTheme() {
+        if (_uiState.value.id < 0) return
+
         viewModelScope.launch {
             try {
                 dynamicThemeRepository.deleteDynamicThemeById(Id(_uiState.value.id))
+                if (isPreferenceState) {
+                    // Reset preferences
+                    userPreferencesRepository.saveIdPreference(-1L)
+                }
             } catch (e: Exception) {
                 val msg = "deleteDynamicTheme: Exception during delete: ${e.message}"
                 Log.e(TAG, msg)
                 onException(msg)
             }
+        }
+    }
+
+    fun setIdPreference(id: Long) {
+        Log.d(APP_TAG, "DynamicThemeViewModel: Setting id preference: $id")
+        viewModelScope.launch {
+            userPreferencesRepository.saveIdPreference(id)
+        }
+    }
+
+    private fun setNamePreference(name: String) {
+        Log.d(APP_TAG, "DynamicThemeViewModel: Setting name preference: $name")
+        viewModelScope.launch {
+            userPreferencesRepository.saveNamePreference(name)
+        }
+    }
+
+    private fun setSourceColorPreference(color: Int, schemeVariant: Variant) {
+        Log.d(APP_TAG, "DynamicThemeViewModel: Setting source color preference: $color")
+        viewModelScope.launch {
+            userPreferencesRepository.saveSourceColorPreference(color, schemeVariant)
         }
     }
 
@@ -172,7 +239,7 @@ open class DynamicThemeViewModel(
 }
 
 data class DynamicThemeUiState(
-    val id: Long = -1,
+    val id: Long = -1L,
     val name: String = "",
     val sourceColorArgb: Int = 0,
     val uiColorSchemeVariant: UiColorSchemeVariant = UiColorSchemeVariant.TonalSpot,
@@ -201,24 +268,37 @@ data class DynamicThemeUiState(
         timestamp = LocalDateTime.now()
     )
 
+    @Composable
+    fun toColorScheme(): ColorScheme = if (isSystemInDarkTheme()) {
+        fromColorSchemeState(darkColorSchemeState)
+    } else {
+        fromColorSchemeState(lightColorSchemeState)
+    }
+
     fun isHorizontalLayout(): Boolean = windowWidthSizeClass != WindowWidthSizeClass.Compact
 
     companion object {
-        fun fromDynamicTheme(dynamicTheme: DynamicTheme): DynamicThemeUiState =
-                createDynamicThemeUiState(
-                    id = dynamicTheme.id,
-                    name = dynamicTheme.name,
-                    sourceColorArgb = dynamicTheme.sourceArgb,
-                    uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(dynamicTheme.colorSchemeVariant)
-                )
+        fun fromDynamicTheme(
+            dynamicTheme: DynamicTheme,
+            windowWidthSizeClass: WindowWidthSizeClass = WindowWidthSizeClass.Compact,
+        ): DynamicThemeUiState = createDynamicThemeUiState(
+            id = dynamicTheme.id,
+            name = dynamicTheme.name,
+            sourceColorArgb = dynamicTheme.sourceArgb,
+            uiColorSchemeVariant = UiColorSchemeVariant.fromVariant(dynamicTheme.colorSchemeVariant),
+            windowWidthSizeClass = windowWidthSizeClass
+        )
 
         fun createDynamicThemeUiState(
             sourceColorArgb: Int,
-            id: Long = -1,
+            windowWidthSizeClass: WindowWidthSizeClass,
+            id: Long = -1L,
             name: String = "",
             uiColorSchemeVariant: UiColorSchemeVariant = UiColorSchemeVariant.TonalSpot,
         ): DynamicThemeUiState {
-            Log.d(TAG, "Create State: Color $sourceColorArgb, Theme $uiColorSchemeVariant")
+            Log.d(
+                TAG, "Create State: $id - $name Color $sourceColorArgb, Theme $uiColorSchemeVariant"
+            )
             val schemeVariant = Variant.entries[uiColorSchemeVariant.ordinal]
             return DynamicThemeUiState(
                 id = id,
@@ -231,6 +311,7 @@ data class DynamicThemeUiState(
                 darkColorSchemeState = ColorExtractor.createDynamicColorScheme(
                     sourceArgb = sourceColorArgb, schemeVariant = schemeVariant, isDark = true
                 ).toColorSchemeState(),
+                windowWidthSizeClass = windowWidthSizeClass
             )
         }
     }
